use non-integer handles (except for figures)
git-svn-id: svn://svn.code.sf.net/p/octave/code/trunk/octave-forge@4385 fc592f45-0881-4044-91ff-c96495a54512

+++ extra/jhandles/src/org/octave/graphics/AxesContainer.java
@@ -40,7 +40,7 @@ public abstract class AxesContainer extends HandleObject
 
 	// Constructor
 
-	public AxesContainer(HandleObject parent, String type, int handle)
+	public AxesContainer(HandleObject parent, String type, double handle)
 	{
 		super(parent, handle, type);
 	}
